@@ -6,7 +6,7 @@ publish_to: "none"
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.19.6 <3.0.0"
+  sdk: ">=3.2.0 <4.0.0"
 
 dependencies:
   flutter:
@@ -25,10 +25,10 @@ dependencies:
   video_player: ^2.7.2
   photo_view: ^0.14.0
   image_picker: ^1.0.4
-  file_picker: ^5.3.1
+  file_picker: ^6.1.1
   carousel_slider: ^4.2.1
   flutter_animate: ^4.2.0
-  http: ^0.13.6
+  http: ^1.1.2
   animations: ^2.0.7
   url_launcher: any
   dots_indicator: any
@@ -36,20 +36,20 @@ dependencies:
   flutter_rating_bar: ^4.0.1
   # flutter_chat_ui: ^0.0.2
   open_filex: ^4.3.4
-  flutter_chat_types: ^3.6.2
+  # flutter_chat_types: ^3.6.2
   uuid: ^4.1.0
   google_maps_flutter: ^2.3.0
   geocoding: ^2.0.5
-  geolocator: ^9.0.2
-  location: ^4.4.0
-  permission_handler: ^10.2.0
+  geolocator: ^10.1.0
+  location: ^5.0.3
+  permission_handler: ^11.1.0
   flutter_staggered_animations: ^1.1.1
   animated_button: ^0.2.0
   flutter_animated_button: ^2.0.3
-  date_field: ^3.0.5
+  # date_field: ^3.0.5
   badges: ^3.1.1
-  flutter_screen_lock: ^9.0.1
-  flutter_otp_text_field: ^1.1.1
+  # flutter_screen_lock: ^9.0.1
+  # flutter_otp_text_field: ^1.1.1
   lottie: ^2.3.2
   # flutter_cupertino_date_picker: ^1.0.26+2
   # flutter_svg: ^2.0.5
@@ -63,7 +63,7 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  flutter_lints: ^2.0.0
+  flutter_lints: ^3.0.1
 
 flutter_launcher_icons:
   android: "launcher_icon"
